user authentication with jwt added

+++ controllers/expensesController.js
@@ -1,15 +1,41 @@
 const Expenses = require('../models/expense');
 
+const User = require('../models/user');
+const jwt = require('jsonwebtoken');
+
 exports.getAllExpenses = (req, res, next) => {
-    Expenses.findAll()
-        .then(response => res.json(response))
-        .catch(err => console.log(err));
+    const userId = jwt.verify(req.query.token, 'ZindagiNaMilegiDubara').id
+    if (userId) {
+        Expenses.findAll({
+            where: {
+                userId: userId
+            }
+        })
+            .then(response => res.json(response))
+            .catch(err => console.log(err));
+    } else {
+        res.status(401).json({
+            message: 'Unauthorized User'
+        });
+    }
 }
 
 exports.getOneExpense = (req, res, next) => {
-    Expenses.findByPk(req.params.id)
-        .then(response => res.json(response))
-        .catch(err => console.log(err));
+    const userId = jwt.verify(req.query.token, 'ZindagiNaMilegiDubara').id
+    if (userId) {
+        Expenses.findByPk({
+            where: {
+                id: req.params.id,
+                userId: jwt.verify(req.query.token, 'ZindagiNaMilegiDubara').id
+            }
+        })
+            .then(response => res.json(response))
+            .catch(err => console.log(err));
+    } else {
+        res.status(401).json({
+            message: 'Unauthorized User'
+        });
+    }
 }
 
 exports.postAddExpense = (req, res, next) => {
@@ -19,22 +45,26 @@ exports.postAddExpense = (req, res, next) => {
     Expenses.create({
         amount: amount,
         description: description,
-        category: category
+        category: category,
+        userId: jwt.verify(req.body.token, 'ZindagiNaMilegiDubara').id
     })
-        .then(response => res.json(response))
+        .then(response => {
+            console.log("5");
+            res.json(response)
+        })
         .catch(err => console.log(err));
 }
 
 exports.putUpdateExpense = (req, res, next) => {
     const expenseId = req.params.id;
-    console.log(req.body);
     Expenses.update({
         amount: parseInt(req.body.amount),
         description: req.body.description,
         category: req.body.category
     }, {
         where: {
-            id: expenseId
+            id: expenseId,
+            userId: jwt.verify(req.body.token, 'ZindagiNaMilegiDubara').id
         }
     }
     )
@@ -43,9 +73,19 @@ exports.putUpdateExpense = (req, res, next) => {
 }
 
 exports.deleteExpense = (req, res, next) => {
-    Expenses.destroy({
-        where: {
-            id: req.params.id
-        }
-    })
+    const userId = jwt.verify(req.query.token, 'ZindagiNaMilegiDubara').id
+    if (userId) {
+        Expenses.destroy({
+            where: {
+                id: req.params.id,
+                userId: jwt.verify(req.query.token, 'ZindagiNaMilegiDubara').id
+            }
+        })
+            .then(response => res.json(response))
+            .catch(err => console.log(err));
+    } else {
+        res.status(401).json({
+            message: 'Unauthorized User'
+        });
+    }
 }
